tend: hoist CANONICAL_SHAPES into substrate services
The canonical cross-modal recipe shapes were authored in
scripts/intern_modality_blueprints.py and a stale subset (7 of the 13
entries) was duplicated as _CANONICAL_SHAPE_NAMES in
api/app/routers/substrate.py. After the list grew from 7 to 13 in
PR #1958 the router silently kept reporting only the original seven
through /canonical_families and /modality_for.

Single source-of-truth now lives in
api/app/services/substrate/modality_shapes — both the intern script and
the router import CANONICAL_SHAPES / canonical_shape_names() from there.
The /canonical_families endpoint now surfaces all 13 interned families
(R_FieldHoldingPresence, R_GroundingMove, R_SequentialScan,
R_KnownStateRecall, R_SuperpositionHold, R_WitnessWithoutIntervention
were previously invisible through the MCP tool surface).

The intern script becomes a thin CLI wrapper; backward-compat aliases
keep test_modality_blueprints.py and test_cross_modal_tools.py importing
through the old names. test_canonical_families_lists_seven_shapes is
re-tuned to assert against canonical_shape_names() so the test grows
with the list.
